+++ unicodeplots/utils/__init__.py
@@ -1,5 +1,4 @@
 from unicodeplots.utils.colors import INVALID_COLOR, Color, ColorType
-from unicodeplots.utils.params import BoxParams, CanvasParams
 from unicodeplots.utils.timing import time_execution
 
-__all__ = ["ColorType", "INVALID_COLOR", "CanvasParams", "BoxParams", "time_execution", "Color"]
+__all__ = ["ColorType", "INVALID_COLOR", "time_execution", "Color"]
